Silence cancel urb errors during iso stream cancel
When cancelling an active iso stream some iso packets will likely complete
while we are cancelling, but we will see them as uncomplete as long as we've
not asked libusb to check for complete transfers, this is inherently racy and
there is nothing we can do. Thus we will sometimes try to cancel already
completed transfers which will fail.

This is not a problem, but libusb spews ugly error messages to stderr when
this happens. This patch silences these error messages.