usbredirhost: Don't attach drivers when releasing interfaces for set_config
When changing the active configuration we must first release all interfaces,
this is done by the usbredirhost_release() helper function, but before
this patch that function also re-attached the in kernel driver, making
the interfaces busy again and causing the set_config call to fail.

This patch fixes this by not doing the kernel driver re-attach when
releasing the interfaces for a set_config call.

Signed-off-by: Hans de Goede <hdegoede@redhat.com>