usbredirhost: Fix handling of iso stream stalls
Before this patch the code would cancel the iso stream, but keep the transfers,
and then on a successfull clear of the stall re-submit the same transfers.
This means however that the transfers would get resubmitted before their
completion callback from the cancel got to run. And re-submitting an
uncompleted transfer will fail.

So instead cancel and free the stream (getting rid of the single user
of cancel_iso_stream which did not want it to also free the urbs) and
when the stall is successfully cleared, alloc a new stream and, if it is
an input stream, submit its transfers.

Signed-off-by: Hans de Goede <hdegoede@redhat.com>