Compile with -Wall -Werror by default and fix existing warnings
The direct trigger for making this change is the fact that the 0.4 release
is broken due to a return without a value in a non void function. Somehow
this happened to always do the right thing in my testing, but not for others.

This is something which should have been caught by a compiler warning, but
I always build from git, and then no -Wall (let alone -Werror) gets passed
to the compiler, which allowed this to slip through. So to avoid future
embarrassment this patch changes the configure script to add various
compiler options to CFLAGS by default, which should catch most detectable
mistakes.

Note that this change modifies CFLAGS directly, so if you're somehow specifying
your own CFLAGS nothing will change (this is intentional). Also all added flags
are checked, so if you've a compiler which is not gcc nothing will change.

This patch also fixes the few warnings actually present in the code, so as to
not have a state in the commit tree where the code does not compile.

Signed-off-by: Hans de Goede <hdegoede@redhat.com>