Rename libusbredirparser.pc to libusbredirparser-0.5.pc
The usbredir 0.5 release introduced the new API for 64 bit packet ids, but
it kept the libusbredirparser.pc name as is, meaning that older versions of
qemu will still have their pkg-config check for usbredirparser fulfilled,
and build with the usb-redir device. Due to the API change there will be
some compiler warnings, but the build will succeed, however the usb-redir
device will be broken on 32 bit machines.

To solve this, this patch renames the libusbredirparser.pc file to
libusbredirparser-0.5.pc, so that it will no longer fulfill the pkg-config
check of the qemu-1.2 and older releases, stopping the (silent) breakage.

A patch has been send to qemu-devel to adjusts qemu master's configure.

Signed-off-by: Hans de Goede <hdegoede@redhat.com>